split DECODE class into mechanisms so programming team can split up work easier.
